# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: >-
    Aircraft Engine MRO Decision Support API. Ingests engine condition data,
    reasons over it with a domain ontology, knowledge graph and rules engine,
    and produces traceable MRO Work Recommendations for human planners and for
    SAP S/4HANA Cloud (mock mode).
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommendations:
    get:
      operationId: listRecommendations
      tags: [recommendations]
      summary: List MRO Work Recommendations
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/RecommendationStatus"
        - name: engineId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
  /recommendations/{id}:
    get:
      operationId: getRecommendation
      tags: [recommendations]
      summary: Get one recommendation with full traceability
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateRecommendation
      tags: [recommendations]
      summary: Edit a recommendation
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "404":
          $ref: "#/components/responses/NotFound"
  /recommendations/{id}/approve:
    post:
      operationId: approveRecommendation
      tags: [recommendations]
      summary: Approve a recommendation
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "404":
          $ref: "#/components/responses/NotFound"
  /recommendations/{id}/reject:
    post:
      operationId: rejectRecommendation
      tags: [recommendations]
      summary: Reject a recommendation
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "404":
          $ref: "#/components/responses/NotFound"
  /recommendations/{id}/push-to-sap:
    post:
      operationId: pushRecommendationToSap
      tags: [recommendations, sap]
      summary: Push an approved recommendation to SAP as a Maintenance Notification
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: SAP notification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SapNotification"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /ontology/classes:
    post:
      operationId: createOntologyClass
      tags: [ontology]
      summary: Add a class to the draft ontology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OntologyClassInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OntologyClass"
        "400":
          $ref: "#/components/responses/BadRequest"
  /ontology/classes/{id}:
    patch:
      operationId: updateOntologyClass
      tags: [ontology]
      summary: Edit or deprecate a class in the draft ontology
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OntologyClassUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OntologyClass"
        "404":
          $ref: "#/components/responses/NotFound"
  /ontology/relationships:
    post:
      operationId: createOntologyRelationship
      tags: [ontology]
      summary: Add a relationship to the draft ontology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OntologyRelationshipInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OntologyRelationship"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /ontology/publish:
    post:
      operationId: publishOntology
      tags: [ontology]
      summary: Validate and publish the draft ontology as a new version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OntologyPublishInput"
      responses:
        "200":
          description: Published version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OntologyVersion"
        "400":
          $ref: "#/components/responses/BadRequest"
  /graph:
    get:
      operationId: getGraph
      tags: [graph]
      summary: Get knowledge-graph nodes and edges
      parameters:
        - name: engineId
          in: query
          required: false
          schema:
            type: string
          description: Scope the graph to one engine's subgraph
        - name: type
          in: query
          required: false
          schema:
            type: string
          description: Filter nodes by class/type
      responses:
        "200":
          description: Graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphData"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RecommendationStatus:
      type: string
      enum: [pending, approved, rejected, pushed, failed]
    Priority:
      type: string
      enum: [routine, expedite, urgent, aog]
...
    Evidence:
      type: object
      required: [parameter, label, value, unit, cycle, timestamp, description]
      properties:
        parameter: { type: string }
        label: { type: string }
        value: { type: number }
        unit: { type: string }
        cycle: { type: integer }
        timestamp: { type: string }
        threshold: { type: ["number", "null"] }
        description: { type: string }
    MaintenanceTask:
      type: object
      required: [ataCode, description, skill, estimatedHours]
      properties:
        ataCode: { type: string, description: "ATA chapter / task code" }
        s1000dCode: { type: ["string", "null"], description: "S1000D / iSpec 2200 task code" }
        description: { type: string }
        skill: { type: string }
        estimatedHours: { type: number }
    LifeLimitedPart:
      type: object
      required: [partNumber, description]
      properties:
        partNumber: { type: string }
        description: { type: string }
        cyclesRemaining: { type: ["integer", "null"] }
    Recommendation:
      type: object
      required:
        - id
        - engineId
        - engineModel
        - tailNumber
        - component
        - failureMode
        - faultDescription
        - priority
        - severity
        - confidence
        - status
        - tasks
        - requiredSkills
        - estimatedDurationHours
        - turnaroundDays
        - workscopeLevel
        - affectedModules
        - evidence
        - regulatoryRefs
        - ruleId
        - ruleName
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        engineId: { type: string }
        engineModel: { type: string }
        tailNumber: { type: string }
        component: { type: string }
        failureMode: { type: string }
        faultDescription: { type: string }
        priority: { $ref: "#/components/schemas/Priority" }
        severity: { type: number, description: "0-1 severity score" }
        confidence: { type: number, description: "0-1 confidence score" }
        status: { $ref: "#/components/schemas/RecommendationStatus" }
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/MaintenanceTask"
        requiredSkills:
          type: array
          items: { type: string }
        estimatedDurationHours: { type: number }
        turnaroundDays: { type: integer }
        workscopeLevel: { type: string }
        affectedModules:
          type: array
          items: { type: string }
        lifeLimitedParts:
          type: array
          items:
            $ref: "#/components/schemas/LifeLimitedPart"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
        regulatoryRefs:
          type: array
          items: { type: string }
        ruleId: { type: string }
        ruleName: { type: string }
        recommendedInductionDate: { type: ["string", "null"] }
        recommendedCompletionDate: { type: ["string", "null"] }
        sapNotificationNumber: { type: ["string", "null"] }
        reviewedBy: { type: ["string", "null"] }
        reviewNotes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    RecommendationUpdate:
      type: object
      properties:
        priority: { $ref: "#/components/schemas/Priority" }
        reviewNotes: { type: string }
        estimatedDurationHours: { type: number }
        turnaroundDays: { type: integer }
        recommendedInductionDate: { type: string }
        recommendedCompletionDate: { type: string }
    ReviewInput:
      type: object
      properties:
        notes: { type: string }
        reviewedBy: { type: string }
...
    OntologyAttribute:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        type: { type: string }
        description: { type: ["string", "null"] }
    OntologyClass:
      type: object
      required: [id, label, description, deprecated, attributes, instanceCount, ruleCount]
      properties:
        id: { type: string }
        label: { type: string }
        description: { type: string }
        parentClass: { type: ["string", "null"] }
        deprecated: { type: boolean }
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/OntologyAttribute"
        instanceCount: { type: integer }
        ruleCount: { type: integer }
    OntologyClassInput:
      type: object
      required: [label, description]
      properties:
        label: { type: string, minLength: 1 }
        description: { type: string }
        parentClass: { type: string }
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/OntologyAttribute"
    OntologyClassUpdate:
      type: object
      properties:
        label: { type: string }
        description: { type: string }
        parentClass: { type: string }
        deprecated: { type: boolean }
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/OntologyAttribute"
    OntologyRelationship:
      type: object
      required: [id, label, domain, range, deprecated]
      properties:
        id: { type: string }
        label: { type: string }
        domain: { type: string, description: "Source class id" }
        range: { type: string, description: "Target class id" }
        description: { type: ["string", "null"] }
        deprecated: { type: boolean }
    OntologyRelationshipInput:
      type: object
      required: [label, domain, range]
      properties:
        label: { type: string, minLength: 1 }
        domain: { type: string }
        range: { type: string }
        description: { type: string }
...
    OntologyPublishInput:
      type: object
      required: [note]
      properties:
        note: { type: string, minLength: 1 }
        author: { type: string }
    OntologyVersion:
      type: object
      required: [version, status, note, createdAt]
      properties:
        version: { type: string }
        status:
          type: string
          enum: [draft, published, superseded]
        note: { type: string }
        author: { type: ["string", "null"] }
        createdAt: { type: string }
        classCount: { type: integer }
        relationshipCount: { type: integer }
    GraphNode:
      type: object
      required: [id, type, label, properties]
      properties:
        id: { type: string }
        type: { type: string, description: "Ontology class id" }
        label: { type: string }
        properties:
          type: object
          additionalProperties: true
    GraphEdge:
      type: object
      required: [id, source, target, label]
      properties:
        id: { type: string }
        source: { type: string }
        target: { type: string }
        label: { type: string }
    GraphData:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
...
    SapNotification:
      type: object
      required: [id, recommendationId, status, createdAt, payload]
      properties:
        id: { type: string }
        recommendationId: { type: string }
        notificationNumber: { type: ["string", "null"] }
        status:
          type: string
          enum: [success, failed]
        mode:
          type: string
          enum: [mock, live]
        errorMessage: { type: ["string", "null"] }
        payload:
          type: object
          additionalProperties: true
        createdAt: { type: string }
...